# shared/contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: AgentRoom HTTP API
  version: 0.11.0
  description: Control and command API for AgentRoom clients and bridges.
servers:
  - url: https://try-status.online/api
    description: Production
  - url: http://127.0.0.1:8787
    description: Local development
...
paths:
...
  /v1/rooms:
    get:
      tags: [rooms]
      operationId: listAccountRooms
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Rooms linked to the authenticated account.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AccountRoomMembership" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [rooms]
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 1, maxLength: 100, pattern: "\\S" }
                visibility: { type: string, enum: [private, public], default: private }
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: "\\S"
                  description: Required for guest creation; defaults to the authenticated account display name.
      responses:
        "201":
          description: Room and its owner membership were created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedRoomAccess" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/public-rooms:
    get:
      tags: [rooms]
      operationId: listPublicRooms
      description: Lists active public rooms that can be joined without an invite code.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Public rooms ordered by newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Room" }
  /v1/rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    patch:
      tags: [rooms]
      operationId: updateRoom
      description: Renames a room or changes public/private visibility. Owner only.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: { type: string, minLength: 1, maxLength: 100, pattern: "\\S" }
                visibility: { type: string, enum: [private, public] }
      responses:
        "200":
          description: Updated room.
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: { $ref: "#/components/schemas/Room" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [rooms]
      operationId: dissolveRoom
      description: Soft-deletes a room and immediately revokes every membership. Owner only.
      security: [{ BearerAuth: [] }]
      responses:
        "204":
          description: Room dissolved and memberships revoked.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/rooms/{roomId}/members:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [rooms]
      operationId: listRoomMembers
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Members currently registered in the room.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [rooms]
      operationId: joinRoom
      description: An account bearer token is optional for human joins and links the new membership to that account. Private rooms require inviteCode; public rooms do not.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [displayName, actorType]
              properties:
                inviteCode: { type: string, minLength: 8, maxLength: 100 }
                displayName: { type: string, minLength: 1, maxLength: 64, pattern: "\\S" }
                actorType: { $ref: "#/components/schemas/ActorType" }
                agentProvider:
                  type: string
                  enum: [claude, codex, other]
              allOf:
                - if:
                    properties: { actorType: { const: agent } }
                  then:
                    required: [agentProvider]
                - if:
                    properties:
                      actorType: { enum: [human, terminal] }
                  then:
                    not: { required: [agentProvider] }
      responses:
        "201":
          description: A room membership and access token were created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoomAccess" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/rooms/{roomId}/members/{memberId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - in: path
        name: memberId
        required: true
        schema: { type: string, minLength: 8, maxLength: 80 }
    delete:
      tags: [rooms]
      operationId: removeRoomMember
      description: Removes a member from the room and revokes their access token. The room owner cannot be removed.
      security: [{ BearerAuth: [] }]
      responses:
        "204":
          description: The member was removed and their token revoked.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/rooms/{roomId}/agent-access:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [agent-access]
      operationId: getAgentAccess
      description: Lists the current account's owned Agents, delegated dispatch access, grants it can manage, and collaboration requests involving its Agents.
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Account-scoped Agent access overview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentAccessOverview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/rooms/{roomId}/agents/{agentId}/claim-code:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/AgentId"
    post:
      tags: [agent-access]
      operationId: rotateAgentClaimCode
      description: Issues a new short-lived ownership claim code. Requires the unclaimed Agent's own member token.
      security: [{ BearerAuth: [] }]
      responses:
        "201":
          description: Previous claim codes were invalidated.
          content:
            application/json:
              schema:
                type: object
                required: [agentClaim]
                properties:
                  agentClaim: { $ref: "#/components/schemas/AgentClaim" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/rooms/{roomId}/agents/{agentId}/claim:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/AgentId"
    post:
      tags: [agent-access]
      operationId: claimAgentOwnership
      description: Links an unclaimed room Agent to the authenticated account. The account must already be a human member of the room.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [claimCode]
              properties:
                claimCode: { type: string, minLength: 12, maxLength: 100, writeOnly: true }
      responses:
        "201":
          description: Agent ownership was claimed.
          content:
            application/json:
              schema:
                type: object
                required: [ownership]
                properties:
                  ownership: { $ref: "#/components/schemas/PublicAgentOwnership" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/rooms/{roomId}/agents/{agentId}/grants:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/AgentId"
    post:
      tags: [agent-access]
      operationId: grantAgentToUser
      description: Allows an account-linked human member to dispatch this Agent, including through a structured @ mention. Requires the Agent owner's account session.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [granteeMemberId]
              properties:
                granteeMemberId: { type: string, minLength: 8, maxLength: 80 }
      responses:
        "201":
          description: Dispatch access granted.
          content:
            application/json:
              schema:
                type: object
                required: [grant]
                properties:
                  grant: { $ref: "#/components/schemas/AgentUserGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/rooms/{roomId}/agents/{agentId}/grants/{grantId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/AgentId"
      - in: path
        name: grantId
        required: true
        schema: { type: string, minLength: 8, maxLength: 80 }
    delete:
      tags: [agent-access]
      operationId: revokeAgentUserGrant
      description: Revokes a user's dispatch access. Requires the Agent owner's account session.
      security: [{ BearerAuth: [] }]
      responses:
        "204": { description: Dispatch access revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/rooms/{roomId}/connector:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [rooms]
      operationId: getRoomConnectorInfo
      description: Returns non-secret AgentRoom CLI metadata for the room owner.
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Copyable CLI command and supported local agent providers.
          content:
            application/json:
              schema:
                type: object
                required: [connectorCommand, connector]
                properties:
                  connectorCommand: { type: string }
                  connector: { $ref: "#/components/schemas/RoomConnectorInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Account session (`ars_`) or room-scoped member (`art_`) token, as allowed by the operation.
...
  schemas:
...
    Room:
      type: object
      required: [id, name, visibility, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        visibility: { type: string, enum: [private, public] }
        createdAt: { type: string, format: date-time }
...
    AccountRoomMembership:
      type: object
      required: [room, member]
      properties:
        room: { $ref: "#/components/schemas/Room" }
        member: { $ref: "#/components/schemas/Member" }
...
    CreatedRoomAccess:
      allOf:
        - $ref: "#/components/schemas/RoomAccess"
        - type: object
          required: [inviteCode, connectorCommand, connector]
          properties:
            inviteCode:
              type: string
              readOnly: true
              description: Raw room invite capability.
            connectorCommand: { type: string }
            connector: { $ref: "#/components/schemas/RoomConnectorInfo" }
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Account or room-member authentication failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...